backend/requirements.txt
fastapi>=0.109.1,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-multipart>=0.0.18,<1.0.0
torch>=2.2.0,<3.0.0 --index-url https://download.pytorch.org/whl/cpu
torchvision>=0.17.0,<1.0.0 --index-url https://download.pytorch.org/whl/cpu
pillow>=10.3.0,<12.0.0
numpy>=1.26.3,<2.0.0
pydantic>=2.5.3,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
httpx>=0.26.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0